Event streaming works again
There's a minor logic change: for CommandStarted event, instead of
"what command was run according to the exe unit" we have "what command
was been run according to our local assessment of processing stage".

This **should** always be the same, assuming no events are missing.

I think we assume this in an implicit way in multiple places, I don't see
a reason why CommandStarted should be different (and it simplifies some
things).