Tests for DecreaseScoreForUnconfirmedAgreement
Test for the "pure" strategy works.

Test for the integration with the Engine doesn't work - but it looks like
the problem is somewhere between current mocking method and the AsyncWrapper,
not in the "important" part of the code.

Also: minor changes in `test_default_strategies`, so now the same helper
function is used in all strategy tests.

Also: Engine.emit() split into two functions.